Fix the locmax issue
