+++ lib/src/widgets/grant_permission_form.dart
@@ -53,6 +53,8 @@ import 'package:solidui/src/utils/webid_message.dart' show webIdCheckMessage;
 import 'package:solidui/src/widgets/grant_permission_dialogs.dart';
 import 'package:solidui/src/widgets/grant_permission_helpers_ui.dart';
 import 'package:solidui/src/widgets/grant_permission_notify.dart';
+import 'package:solidui/src/widgets/grant_permission_webid_utils.dart'
+    show isSelfShare, selfShareMessage;
 import 'package:solidui/src/widgets/group_webid_input.dart';
 import 'package:solidui/src/widgets/ind_webid_input.dart'
     show indWebIdFormatError;
@@ -284,8 +286,8 @@ class _GrantPermissionFormState extends State<GrantPermissionForm> {
     // third-party agent), so catch it here with a friendly explanation
     // rather than surfacing the raw exception via the generic failure
     // snackbar.
-    if (_isSelfShare(webId)) {
-      await _alert(_selfShareMessage);
+    if (isSelfShare(webId, widget.ownerWebId)) {
+      await _alert(selfShareMessage);
       return false;
     }
     final result = await validateWebId(webId);
@@ -300,41 +302,6 @@ class _GrantPermissionFormState extends State<GrantPermissionForm> {
       finalWebIdList = [webId];
     });
     return true;
-  }
-
-  /// Message shown when the recipient WebID is the same as the resource
-  /// owner. Surfaced both for the individual recipient flow and as part of
-  /// the group flow when one of the group entries matches the owner.
-  static const String _selfShareMessage =
-      'This resource is owned by you, so you already have full access. Please '
-      'enter the WebID of another user if you would like to share this '
-      'resource.';
-
-  /// Returns true when [webId] refers to the same WebID as the resource
-  /// owner, ignoring surrounding whitespace and trailing slashes. The
-  /// comparison is intentionally loose because the owner's WebID is
-  /// captured from different sources (login session, ACL lookup) and the
-  /// trailing-slash form may vary between them.
-  bool _isSelfShare(String webId) {
-    final entered = _normaliseWebId(webId);
-    if (entered.isEmpty) return false;
-    final owner = _normaliseWebId(widget.ownerWebId);
-    return owner.isNotEmpty && entered == owner;
-  }
-
-  String _normaliseWebId(String webId) {
-    final trimmed = webId.trim();
-    if (trimmed.isEmpty) return trimmed;
-    // Drop a trailing slash on the WebID document portion so that, for
-    // example, `https://alice.example/profile/card#me` and
-    // `https://alice.example/profile/card/#me` compare equal.
-    final hashIndex = trimmed.indexOf('#');
-    final docPart = hashIndex >= 0 ? trimmed.substring(0, hashIndex) : trimmed;
-    final fragment = hashIndex >= 0 ? trimmed.substring(hashIndex) : '';
-    final canonicalDoc = docPart.endsWith('/')
-        ? docPart.substring(0, docPart.length - 1)
-        : docPart;
-    return '$canonicalDoc$fragment'.toLowerCase();
   }
 
   /// Validate the group fields typed by the user and, when valid,
@@ -362,8 +329,8 @@ class _GrantPermissionFormState extends State<GrantPermissionForm> {
     // owner. The ACL layer rejects the owner appearing as a third-party
     // agent, so flag it here with a friendly message instead of letting
     // the request fail downstream with a generic snackbar.
-    if (webIdList.any(_isSelfShare)) {
-      await _alert(_selfShareMessage);
+    if (webIdList.any((webId) => isSelfShare(webId, widget.ownerWebId))) {
+      await _alert(selfShareMessage);
       return false;
     }
 
@@ -554,37 +521,8 @@ class _GrantPermissionFormState extends State<GrantPermissionForm> {
             // Capture the ScaffoldMessenger now, while the dialog and its host
             // page are still mounted, so the success feedback can be shown
             // after this dialog is popped.
-            final messenger = ScaffoldMessenger.of(context);
-            void showSnack(
-              String message,
-              Color backgroundColor, {
-              Duration duration = const Duration(seconds: 4),
-            }) {
-              // Showing a SnackBar requires a Scaffold registered with the
-              // messenger. Depending on how the host embeds this form, and on
-              // the exact moment the dialog is dismissed, the messenger can
-              // momentarily have no Scaffold, which throws the
-              // "_scaffolds.isNotEmpty" assertion. A confirmation toast is
-              // non-critical (the grant has already succeeded), so we guard the
-              // call and retry once on the next frame rather than ever letting
-              // it crash the app.
-              SnackBar build() => SnackBar(
-                    content: Text(message),
-                    backgroundColor: backgroundColor,
-                    duration: duration,
-                  );
-              try {
-                messenger.showSnackBar(build());
-              } on Object catch (_) {
-                WidgetsBinding.instance.addPostFrameCallback((_) {
-                  try {
-                    messenger.showSnackBar(build());
-                  } on Object catch (e) {
-                    debugPrint('Could not show snackbar "$message": $e');
-                  }
-                });
-              }
-            }
+            final showSnack =
+                makeResilientSnackBar(ScaffoldMessenger.of(context));
 
             // Grant permission for each resource sequentially. When
             // resourceNames is provided all resources share the same
